bug fixing dbgit single ,add displayAlert to content slice

+++ src/global/git/models/shadow-field-transformer/fileTransform.js
@@ -6,12 +6,23 @@ export const fileTransform = {
 		return await self.git.getFile64Data(oValue)
 	},
 	async transformOnSave(oValue, oRow, self) {
-		const fileInfo = getFileInfo(oValue, true)
-		console.log(fileInfo)
+		let fileInfo
+		try{ 
+			fileInfo = getFileInfo(oValue, true)
+		}catch(e){
+			console.log(`fileTransform error: getFileInfo failed`)
+			return null
+		}
+		// console.log(fileInfo)
+
 		const oldPath = self.shadowedFieldValues[oRow[self.pk]].image
-		const oldGitPath = oldPath.replace(/^\//, "")
+		const oldGitPath = oldPath ? oldPath.replace(/^\//, "") : null
 		const newGitPath = `${self.imageUploadPath}/${fileInfo.name}`
 		const newPath = `/${newGitPath}`
+
+		if(fileInfo.name === 'DEFAULT_DATA_URL_FILENAME'){
+			return oldPath?oldPath:null
+		}
 		console.log({ oldPath, newPath, newGitPath })
 		if (oldPath !== newPath) {
 			self.commitQueues.push(newGitPath)
@@ -30,13 +41,14 @@ export const fileTransform = {
 			}
 
 			console.log(`Do delete old file ${oldFsPath}`)
-			try {
-				await self.fs.unlinkSync(oldFsPath)
-				await self.git.remove(oldGitPath)
-				self.commitMessageQueues.push(`Delete ${oldGitPath}`)
-			} catch (e) {
-				console.log(`lfs: cant delete ${oldPath}`, e)
-			}
+			if(oldPath){
+						try {
+							await self.fs.unlinkSync(oldFsPath)
+							await self.git.remove(oldGitPath)
+							self.commitMessageQueues.push(`Delete ${oldGitPath}`)
+						} catch (e) {
+							console.log(`lfs: cant delete ${oldPath}`, e)
+						}}
 		} else {
 			console.log(`skip oldPath === newPath is true`)
 		}
